Implement interface for game statistics
