+++ src/logics/game_statistics/GameStatistics.java
@@ -2,32 +2,35 @@ package logics.game_statistics;
 
 import utilities.counter.CounterImpl;
 
-public class GameStatistics {
-	private CounterImpl killedEnemies;
-	private CounterImpl passedRooms;
-	private CounterImpl collectedArtefacts;
-
-	public void increaseKilledEnemies() {
-		killedEnemies.increment();
-	}
-
-	public void increasePassedRooms() {
-		passedRooms.increment();
-	}
+/**
+ * 
+ * Contains game's statistics.
+ *
+ */
+public interface GameStatistics {
+	/**
+	 * Increase killed enemies counter.
+	 */
+	public void increaseKilledEnemies();
+	/**
+	 * Increase passed rooms counter.
+	 */
+	public void increasePassedRooms();
+	/**
+	 * Increase collected artefacts counter.
+	 */
+	public void increaseCollectedArtefacts();
+	/**
+	 * @return killed enemies counter.
+	 */
+	public CounterImpl getKilledEnemies();
+	/**
+	 * @return passed rooms counter.
+	 */
+	public CounterImpl getPassedRoom();
+	/**
+	 * @return collected artefacts counter.
+	 */
+	public CounterImpl getCollectedArtefacts();
 	
-	public void collectedArtefacts() {
-		passedRooms.increment();
-	}
-
-	public CounterImpl getKilledEnemies() {
-		return killedEnemies;
-	}
-
-	public CounterImpl getPassedRoom() {
-		return passedRooms;
-	}
-	
-	public CounterImpl getCollectedArtefacts() {
-		return collectedArtefacts;
-	}
 }
